requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
google-generativeai==0.8.5
pydantic-settings==2.8.1
email-validator==2.2.0
